requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
djangorestframework==3.16.0
pika==1.3.2
